\name{LCAtoolbox-package}
\alias{LCAtoolbox-package}
\alias{LCAtoolbox}
\docType{package}
\title{
\packageTitle{LCAtoolbox}
}
\description{
\packageDescription{LCAtoolbox}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LCAtoolbox}
\packageIndices{LCAtoolbox}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LCAtoolbox}

Maintainer: \packageMaintainer{LCAtoolbox}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
