% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.SLTCA.R
\name{predict.SLTCA}
\alias{predict.SLTCA}
\title{predict.SLTCA: Predict membership probabilities from a fitted SLTCA model}
\usage{
\method{predict}{SLTCA}(object, newdata, target.time = Inf, ...)
}
\arguments{
\item{object}{Fitted SLTCA object.}

\item{newdata}{New data used for prediction.}

\item{target.time}{Time at which the prediction should be performed; only
observations with time <= target.time are used. Defaults to Inf (use all).}

\item{...}{Not used.}
}
\value{
A list with tau (matrix of posterior membership probabilities, one
  row per subject), modal (modal class assignment) and data (the subset of
  newdata used for prediction).
}
\description{
Computes posterior latent class membership probabilities for
  (new) subjects using their longitudinal observations up to a target time.
  newdata must be in the same long format as the data used to fit the model
  (columns id, time, num_obs, the outcomes and covariates). Outcomes may be
  missing (NA) at any subset of visits.
}
\examples{

dat <- simulation.SLTCA(500)
res <- SLTCA(dat,num_class=2,covx="baselinecov",vary=paste("y.",1:6,sep=''),covgee="time",
             Y_dist=c('poi','poi','bin','bin','normal','normal'),varest=TRUE,verbose=TRUE,stop.rule="tau")
pred <- predict(res,newdata=dat,target.time=1)

}
\references{
Hart, K. R., Fei, T., & Hanfelt, J. J. (2021). Scalable and robust latent trajectory class analysis using artificial likelihood. Biometrics, 77(3), 1118-1128.
}
\author{
Teng Fei. Email: feit1@mskcc.org
}
