% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{simulation.lcphR}
\alias{simulation.lcphR}
\title{simulation.lcphR: simulate a dataset for latent class survival analysis}
\usage{
simulation.lcphR(n = 1000)
}
\arguments{
\item{n}{sample size of the simulated dataset.}
}
\description{
Simulate a dataset according to the scenario (I) in the manuscript.
}
\examples{

dat <- simulation.lcphR(n=1000)

}
\references{
Fei, T., Hanfelt, J., & Peng, L. (2022). Latent Class Analysis with Semi-parametric Proportional Hazards Submodel for Time-to-event Data. arXiv preprint arXiv:2202.00775.
}
\author{
Teng Fei. Email: feit1@mskcc.org
}
