% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLTCA.R
\name{SLTCA}
\alias{SLTCA}
\title{SLTCA: Scalable and Robust Latent Trajectory Class Analysis Using Artificial Likelihood}
\usage{
SLTCA(
  dat,
  num_class,
  covx,
  vary,
  covgee = NULL,
  Y_dist,
  tolEM = 0.001,
  maxiterEM = 500,
  initial = "null",
  cor = "ar1",
  init.tau = NULL,
  varest = FALSE,
  lbound = 8,
  verbose = FALSE,
  stop.rule = "PAR",
  keep.switch = FALSE
)
}
\arguments{
\item{dat}{Input data.}

\item{num_class}{Number of latent classes.}

\item{covx}{Names of baseline covariates.}

\item{vary}{Names longitudinal outcomes.}

\item{covgee}{Names of covariates in GEE models.}

\item{tolEM}{Stopping criteria for the EM algorithm.}

\item{maxiterEM}{Number of maximum iterations of the EM algorithm.}

\item{initial}{Initialization setting, either 'kmcov', 'null', 'random', 'true' or provided in 'init.tau'.}

\item{cor}{Working correlation structure, currently only supporting "ar1".}

\item{init.tau}{Provided membership probability based on prior knowledge.}

\item{varest}{Variance estimation indicator.}

\item{lbound}{Constant bound of linear projection of approximated likelihood ratio.}

\item{verbose}{Indicating whether output progress.}

\item{stop.rule}{Stopping rule, either "PAR" for parameter L2 norm, or "tau" for posterior probability.}
}
\value{
Returns parameter estimation and variance estimation.
}
\description{
Scalable and Robust Latent Trajectory Class Analysis Using Artificial Likelihood
}
\examples{

dat <- simulation.SLTCA(500)
res <- SLTCA(dat,num_class=2,covx="baselinecov",vary=paste("y.",1:6,sep=''),
             Y_dist=c('poi','poi','bin','bin','normal','normal'),verbose=TRUE,stop.rule="tau")

}
\references{
Fei et al (202x)
}
\author{
Teng Fei. Email: tfei@emory.edu
}
